requirements.txt
PIL==1.1.7
PyYAML==3.10
reportlab==2.7
